chore: split for_head.rs / var_decl_sources.rs out of 2000-line files
