+++ crates/perry-hir/src/destructuring/mod.rs
@@ -28,6 +28,7 @@ mod assignment_stmt;
 mod helpers;
 mod pattern_binding;
 mod var_decl;
+mod var_decl_sources;
 
 pub(crate) use assignment_expr::lower_destructuring_assignment;
 pub(crate) use assignment_stmt::{
