+++ crates/perry-hir/src/lower/mod.rs
@@ -48,11 +48,12 @@ mod unimpl_hints;
 pub(crate) use context::*;
 mod stmt;
 pub(crate) use stmt::*;
+mod for_head;
+pub(crate) use for_head::{for_head_binding_stmts, predefine_for_head, ForHeadBinding};
 mod stmt_loops;
 pub(crate) use stmt_loops::{
-    for_head_binding_stmts, insert_iterator_close_on_abrupt, iterator_close_guarded_stmt,
-    iterator_next_call, lazy_iter_for_stmt, lazy_or_index_elem, lower_stmt_for_in,
-    lower_stmt_for_of, predefine_for_head,
+    insert_iterator_close_on_abrupt, iterator_close_guarded_stmt, iterator_next_call,
+    lazy_iter_for_stmt, lazy_or_index_elem, lower_stmt_for_in, lower_stmt_for_of,
 };
 mod module_decl;
 pub(crate) use module_decl::*;
